Optimizations
Squashed commit of the following:

commit e9be6e552a95abd6335bad3a38a81da5f38d5537
Author: Andy Schott <andy@andyschott.com>
Date:   Mon Oct 13 20:22:08 2025 -0400

    use doubles instead of decimals

commit 92a576a3a1762b71c01aa0f6be518e9d7d4cd25d
Author: Andy Schott <andy@andyschott.com>
Date:   Sun Oct 12 22:52:11 2025 -0400

    no need to compute the inverse of the sphere transform every time

commit f245d912be02f5feeed0d6c55ceabd358f7e97f2
Author: Andy Schott <andy@andyschott.com>
Date:   Sun Oct 12 22:45:41 2025 -0400

    origin can be readonly

commit 9e91d75349f495e2edae83eb9dca5d044572737e
Author: Andy Schott <andy@andyschott.com>
Date:   Sun Oct 12 22:42:53 2025 -0400

    add private origin to Sphere

    it is always set to 0, 0, 0 for now

commit dce54eceb8f2e0c056bb9ecea23298c25e600ece
Author: Andy Schott <andy@andyschott.com>
Date:   Sun Oct 12 22:40:25 2025 -0400

    only need to compute the determinant once when inverting a matrix

commit e35fc5bf28020be8b7de934f842c44e68ca11b7e
Author: Andy Schott <andy@andyschott.com>
Date:   Sun Oct 12 22:36:35 2025 -0400

    improve array copying

commit 99310973b9d38315ee462baf5df72e9a44fd06be
Author: Andy Schott <andy@andyschott.com>
Date:   Sun Oct 12 22:34:41 2025 -0400

    code cleanup

+++ RayTracerChallenge.Tests/Domain/MaterialTests.cs
@@ -14,9 +14,9 @@ public class MaterialTests
         var m = new Material();
         
         Assert.Equal(new Color(1, 1, 1), m.Color);
-        Assert.Equal(0.1M, m.Ambient);
-        Assert.Equal(0.9M, m.Diffuse);
-        Assert.Equal(0.9M, m.Specular);
+        Assert.Equal(0.1, m.Ambient);
+        Assert.Equal(0.9, m.Diffuse);
+        Assert.Equal(0.9, m.Specular);
         Assert.Equal(200, m.Shininess);
     }
 
@@ -30,7 +30,7 @@ public class MaterialTests
         
         var result = _material.Lighting(light, _position,
             eyeVector, normalVector);
-        var expectedResult = new Color(1.9M, 1.9M, 1.9M);
+        var expectedResult = new Color(1.9, 1.9, 1.9);
         
         Assert.Equal(expectedResult, result);
     }
@@ -39,7 +39,7 @@ public class MaterialTests
     public void LightingWithEyeBetweenLightAndSurface_EyeOffset()
     {
         var eyeVector = Tuple.CreateVector(0,
-            (decimal)Math.Sqrt(2)/2, -(decimal)Math.Sqrt(2)/2);
+            Math.Sqrt(2)/2, -Math.Sqrt(2)/2);
         var normalVector = Tuple.CreateVector(0, 0, -1);
         var light = new PointLight(Tuple.CreatePoint(0, 0, -10),
             new Color(1, 1, 1));
@@ -61,7 +61,7 @@ public class MaterialTests
         
         var result = _material.Lighting(light, _position,
             eyeVector, normalVector);
-        var expectedResult = new Color(0.7364M, 0.7364M, 0.7364M);
+        var expectedResult = new Color(0.7364, 0.7364, 0.7364);
         
         Assert.Equal(expectedResult, result);
     }
@@ -70,14 +70,14 @@ public class MaterialTests
     public void LightingWithEyeInPathOfReflection()
     {
         var eyeVector = Tuple.CreateVector(0,
-            -(decimal)Math.Sqrt(2)/2, -(decimal)Math.Sqrt(2)/2);
+            -Math.Sqrt(2)/2, -Math.Sqrt(2)/2);
         var normalVector = Tuple.CreateVector(0, 0, -1);
         var light = new PointLight(Tuple.CreatePoint(0, 10, -10),
             new Color(1, 1, 1));
         
         var result = _material.Lighting(light, _position,
             eyeVector, normalVector);
-        var expectedResult = new Color(1.6364M, 1.6364M, 1.6364M);
+        var expectedResult = new Color(1.6364, 1.6364, 1.6364);
         
         Assert.Equal(expectedResult, result);
     }
@@ -92,7 +92,7 @@ public class MaterialTests
         
         var result = _material.Lighting(light, _position,
             eyeVector, normalVector);
-        var expectedResult = new Color(0.1M, 0.1M, 0.1M);
+        var expectedResult = new Color(0.1, 0.1, 0.1);
         
         Assert.Equal(expectedResult, result);
     }
