+++ RayTracerChallenge/Constants.cs
@@ -2,5 +2,5 @@ namespace RayTracerChallenge;
 
 static class Constants
 {
-    public const decimal Epsilon = 0.0001M;
+    public const double Epsilon = 0.0001;
 }
